python-requirements.txt
pyparallel
pyaudio
